Backend/requirements.txt
pymupdf
google-genai
google-generativeai
python-dotenv
fastapi[standard]
openai
youtube-transcript-api
validators
pydantic
pydantic-core
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
elasticsearch
sentence-transformers
beautifulsoup4
SQLAlchemy
databases
asyncpg
boto3
langchain
langchain-groq
composio-langchain
icalendar
langchain-google-genai